--- offix-offline/src/main/AndroidManifest.xml
@@ -1,20 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="org.aerogear.offixoffline">
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
-    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
-
-    <application>
-        <provider android:exported="false"
-                  android:authorities="org.aerogear.offixoffline"
-                  android:name="org.aerogear.offixoffline.provider.Provider"/>
-        <receiver
-                android:name="org.aerogear.offixoffline.NetworkChangeReceiver"
-                android:exported="true">
-            <intent-filter>
-                <action android:name="android.net.conn.CONNECTIVITY_CHANGE"/>
-            </intent-filter>
-        </receiver>
-        <service android:name=".OfflineSyncService"/>
-    </application>
-</manifest>
