+++ offix/src/main/AndroidManifest.xml
@@ -1,0 +1,20 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="org.aerogear.offix">
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
+
+    <application>
+        <provider android:exported="false"
+                  android:authorities="org.aerogear.offix"
+                  android:name="org.aerogear.offix.provider.Provider"/>
+        <receiver
+                android:name="org.aerogear.offix.NetworkChangeReceiver"
+                android:exported="true">
+            <intent-filter>
+                <action android:name="android.net.conn.CONNECTIVITY_CHANGE"/>
+            </intent-filter>
+        </receiver>
+        <service android:name="org.aerogear.offix.OfflineSyncService"/>
+    </application>
+</manifest>
