+++ offix/src/main/java/org/aerogear/offix/persistence/MutationDao.kt
@@ -1,0 +1,26 @@
+package org.aerogear.offix.persistence
+
+import android.arch.persistence.room.*
+
+@Dao
+interface MutationDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun insertMutation(mutation: Mutation): Long
+
+    @Delete
+    fun deleteMutation(mutation: Mutation)
+
+    @Query("SELECT * FROM MutationOffline")
+    fun getAllMutations(): List<Mutation>
+
+    @Query("SELECT * FROM MutationOffline WHERE sNo =:sno")
+    fun getAMutation(sno: Int): Mutation
+
+
+    @Query("DELETE FROM MutationOffline WHERE sNo= :sno ")
+    fun deleteCurrentMutation(sno: Int)
+
+    @Query("DELETE FROM MutationOffline")
+    fun deleteAllMutations()
+}
