+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':sample', ':offix-offline'
+include ':sample', ':offix'
